requirements.txt
pyserial
sounddevice
